Moved DeployAndWaitForSharingRecalculation to its own file
